+++ tests/test_upload.py
@@ -6,7 +6,7 @@ import pytest
 from moto import mock_aws
 
 from grz_cli.file_operations import calculate_sha256
-from grz_cli.upload import S3BotoUploadWorker
+from grz_cli.workers.upload import S3BotoUploadWorker
 
 
 @pytest.fixture(scope="module")
